requirements.txt
tensorflow
pandas
keras
numpy
opencv-python
matplotlib
pillow
scipy
scikit-learn
streamlit
albumentations
tqdm
torch
torchvision
torchaudio
